pyproject.toml
[tool.poetry]
name = "noodles"
version = "0.1.0"
description = ""
authors = ["Spaceocean <spaceocean@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.13.1"
marshmallow = "^3.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
